Backen terminado

+++ routes/products.js
@@ -3,29 +3,41 @@ var router = express.Router();
 var productModel = require('../models/product');
 
 
-router.put("/:id", function (req, res, next) {
-  res.json("Modificando producto " + req.params.id);
- });
- 
- 
- router.delete("/:id", async function (req, res, next) {
+//Modificar un producto por ID
+router.put("/:id", async function (req, res, next) {
 
+  //Buscar un producto por ID y actualizar sus campos
+if(!productModel.findOne({id: req.params.id}).count()) return res.json({error: "No se encontró el producto con Id " + req.params.id});
+
+  await productModel.updateOne({id: req.params.id}, {$set: {
+    description: req.body.description,
+    price: req.body.price,
+    images: req.body.images
+  }}).exec();
+
+
+  return res.json("Producto " + req.params.id + " modificado exitosamente");
+});
+
+
+//Eliminar un producto por ID
+router.delete("/:id", async function (req, res, next) {
 
   //Buscar un producto por ID y regresa una lista
-const resul = await productModel.find({id: req.params.id}).exec();
+  const resul = await productModel.find({id: req.params.id}).exec();
 
-
-//Si se encontró lo elimina
-if (resul.length > 0) {
-  await productModel.deleteOne({id: req.params.id});
-  res.json("Eliminando producto");
-} else {
-  res.json({error: "No se encontró el producto con Id " + req.params.id})
-}
+  //Si se encontró lo elimina
+  if (resul.length > 0) {
+    await productModel.deleteOne({id: req.params.id});
+    res.json("Eliminando producto");
+  } else {
+    res.json({error: "No se encontró el producto con Id " + req.params.id})
+  }
 });
- 
- router.post("/", async function (req, res, next) {
 
+
+//Agregar un nuevo producto
+router.post("/", async function (req, res, next) {
 
   const product = new productModel({
       id: req.body.id, //Extra el Id pasado por el body
@@ -35,19 +47,17 @@ if (resul.length > 0) {
       images: req.body.images
     });
 
-
-    const result = await product.save(); // Lo guarda en Mongo
-    res.json('Registro Agregado exitosamente');
+  const result = await product.save(); // Lo guarda en Mongo
+  res.json('Registro Agregado exitosamente');
 });
+
 
 //Listado de todos los productos
 router.get("/", async function (req, res, next) {
 
-
   const resultado = await productModel.find();
   res.json(resultado);
 });
-
 
 
 module.exports = router;
